Update issue 55
Casted each pixel's value to a float, then to a Decimal to actually look at what's  being asserted.  THe following error message is produced:

AssertionError: Unequal pixel values detected at row 0 index 805:1.08000004291534423828125 1.0799999237060546875

The first value is the output of InVEST3 carbon model, the second is the output of the InVEST2.1 carbon model.