Fixes issue 65
Carbon economic valuation asserts correctly against the output of invest2.1 out to 4 decimal places.

Also, I found a bug in invest2.1 which is unfortunately not possible to isolate to the test case.  See carbon_core lines 97, 98.

According to the user's manual , the effects of discounted prices and market rate changes are summed up from 0 to n, where n is supposed to be the number of years of the study minus 1 (i.e. lulc_cur date: 2000, lulc_fut date = 2030, years of study = 30, so n should be 29).  InVEST 2.1 calculates the sum of price and market rate changes from 0 to n+1 (they do not subtract 1 from the number of years).

The number of years is used in carbon_core on lines 98 and 103 for different purposes.  On line 98, 1 is subtracted from it, on line 103, it is used outright.  For this reason, you must switch the comments on lines 97, 98 to allow invest3 to produce the correct output raster.

I have left the invest2-ready line, as the valuation test verifies as such.