Update issue 41
Code now correctly burns the feature ID into the output raster.  At the moment, the output raster is just set to be the actual output raster (output from the carbon_storage model), but this can easily be changed to be the raster we created in memory for the rest of the processing.