Update issue 98
This issue (issue 98) was caused by an incorrect 'index' variable being passed to calcFeatureHWP.  The problem is an error in the user's manual, where the opper bounds of the summation for HWP in the 2.1 documentation calls for the term to the (-1) power.  If you refer to the documentation for invest 2.0, this should be a straightforward (-1) rather than a (^(-1)).

Noteworthy changes to carbon_core:
 - This issue has been fixed in carbon_core, pending a careful review of the math.
 - HWP is only calculated for a given feature only if the field 'cut_*' has been defined as anything other than 0.  This saves many iterations of the summation.
 - I removed the call to python's math.floor() in calcFeatureHWP to assist with debugging.