Fixes issue 102
Added logic to invest_carbon_core to only create the files that should be created.  This determination is based on the presence of certain key input files.  Added a 'calc_uncertainty' boolean in the args dict to determine whether to create uncertainty output rasters.  Added assertions to existing invest_test unittests to ensure that only the correct files are created.