Update issue 101
Added missing fields to the toolbox and attempted to add the validation code necessary to properly enable/disable fields.

The validation code seems to only work reliably when altering parameter fields.  Validation is not so reliable if a given field already has information that has been read from our JSON file.  In this latter case, the user would then need to re-enter that field's data to see the proper validation take place.