Update issue 93
Wrote a method in invest_core that takes a shapefile as input and generates an empty raster as output that tightly bounds that shapefile.  This updates issue 93 since this function could be used to generate an output raster from a shapefile generated as the intersection of many bounding boxes on input rasters.