Update issue 150
Added a new class for a processThread, which is launched to validate the user's input.  When the processThread finishes, its finished() signal connects to the startQProcess function.  This function checks to see if the validator returned any errors.  If so, the errors are written to the message area.  Otherwise, a validation completion status message is printed to the status area and the QProcess begins as normal.