Update issue 150
The function showEvent(), if provided, is called when the modelDialog's self.exec_() function is called.  showEvent writes a status message to the messages window and starts the validation thread.

Unfortunately, there appears to be a little overhead on the thread.start() function, which causes a slight delay between when the user presses ok in the main window and when the modal window pops up.  comment out the thread.start() and the lag goes away.  I wonder if there's a way to show the window before starting the thread.