Update issue 155
Replaced existing code in _arc files for carbon_biophysical, carbon_valuation, timber models.

path to carbon_valuation file was not stored as a relative path.  Fixed with this commit.

Added descriptions to the three models.  Unfortunately, there don't appear to be any labels or other 'parameter' features of the parameters area that would allow us to annotate the model.  Instead, the only readily editable text appears to be the help portion of the script window.  Help area seems to strip non-space whitespace, so we have very little options for formatting.