Cast the shape_source uri to a python string.
When a json string is read into a python dictionary by simplejson, strings are saved as unicode strings and not standard python strings.  CopyDataSource() on line 46 of timber.py requires a python string as its second argument, so this cast is necessary for all cases where JSON objects are read from the command line (which is the case for every instance of our UI).


Former-commit-id: 86753929c97f3c375da15642879f9d0f4e0896e9