Update issue 190
Validator code is called properly with this commit.

Note that the time.sleep() call is necessary at the moment because we need to have some sort of mechanism that blocks until the validator terminates.  If time.sleep() is not there, we get some funky errors resulting from the Qprocess being destroyed before it's finished processing.


Former-commit-id: 0de257c842e291269f0de0ab0f68bc4322ecd927