Update issue 190
IUI now calls through to validator and model.  Validator/model dialog flashes briefly to the screen before it is disabled.  Perhaps there is a way to set a local variable to prevent the modal dialog from being shown on exec_()?

iui_main still needs to be cleaned up ... a number of debugging print statements remain.


Former-commit-id: 041736f72d51557f2da93f6ecabd5bf6d5437906