Update issue 172
Added function to DynamicUI that loops through all elements and resets their values to default (if the element had a default).  If the element did not have a default value, the element is left alone.

Because this issue's functionality is no longer preventing the conclusion of issue 190, I'll leave the rest of this issue (the creation of a button to execute this function) for later completion. (Fixes issue 190).


Former-commit-id: 0690c00c86c50b3524cfdbd5fb0cc0da91440b78