Update issue 194
Adding a hyperlink to the top of the IUI window.  This hyperlink (when clicked) links to the model's local documentation URI, located at /invest/documentation/__________.html.  Added an attribute to the JSON file to make this possible and edited the existing JSON files to link to their appropriate docs chapters.


Former-commit-id: 88b5afcf0355ccbbda33a89304ed35ebf7b2953b