Update issue 194
Adding a hyperlink to the top of the IUI window.  This hyperlink (when clicked) links to the model's local documentation URI, located at /invest/documentation/__________.html.  Added an attribute to the JSON file to make this possible and edited the existing JSON files to link to their appropriate docs chapters.


Former-commit-id: 6fe88dbf1a1181179d1d9f4eefd926b02a13e3dc