Update issue 194
Converted the documentation link to a QPushButton in the same position as it was previously.  Also added a feedback button to this same buttonbox (issue 168).  Both callbacks are implemented using qdesktopservices to launch the appropriate URI.  This works fine for the documentation link (since it's a webpage anyways), but it doesn't really work reliably for the mailto: link.  On my windows machine (and perhaps this is because I don't have a default mail client), the button doesn't do anything; on my linux machine, it opens Chromium but only opens a new tab.

EDIT (relevant to issue 168): I just installed Thunderbird and I once I press the little feedback star, thunderbird correctly opens up a new message window.  This takes a second or two, however, so it would probably be a good idea to pop up a modal window saying something like 'launching mail client' or something to that effect.


Former-commit-id: 7ba4cab5fe2a529924e74c9343b48b2e455ca8e3