Rewrote pointShapeToDict and interpolateFields functions. pointShapeToDict was renamed to 'getPointsValues'. interpolateFields was renamed to 'interpPointsOverRaster'
getPointsValues takes the same parameters as the old function but now returns and array of two elements, the first being an array of all the points from the shapefile and the second being an array of all the values corresponding to those points.

interPointsOverRaster now uses the scipy.interpolate.LinearNDInterpolator() function to interpolate the points and values since they really represent unstructured data.


Former-commit-id: cc4a6e842a3ba8588fdc71fff8cf8d67af61732c