Update issue 287
Added a link to system site packages so we can use the ubuntu gdal, numpy, and scipy python packages and not worry about building them ourselves.  This will translate over to the win32 environment by assuming those packages are installed there too.  In the end, we'll ensure this is true by running it in the installer.


Former-commit-id: d157d492a39533dbc9b610029080dd906cfa1f3d