Update issue 314.
Added a shapefile argument to be past to valuation in the first test case.   This argument would represent the shapefile that holds all the attributes such as depth, wave energy, wave power, lat, long, etc...

It is still in question whether this file should be added by user, or called upon in code, since it should have already been created due to biophysical run.


Former-commit-id: d12eddf56a82e6ccbeef5627367185104c6b45c7