Update issue 315
Wrote functions getPoints and calcDist.
getPoints takes a point shapefile and returns a 2D array of lists, where the inner lists are [x,y] (the coordinates).

calcDist takes two numpy arrays of points and calculates the distance, choosing the shortest distance.  It returns a list of the shortest distances and a list of id's corresponding to the distances.  these id's are used for determining which land point was the shortest distance.


Former-commit-id: 9e35e07a601638321bf3e513221fa3680e52b6bc