Update issue 330
Was able to make the raster for W2L.  There are descrepancies between pixel value on raster and what same point says for point on shapefile.  The values are in the same ballpark but a bit off.  May be due to projections.

Still need to make the other raster and determine issues with values.


Former-commit-id: b35c10a12e9851c29eb736597b8bdbdb338b9e45