Update issue 330
Added the creation the 2nd raster L2G.

However, I'm not so sure these rasters should be interpolated, at least not L2G, as it is either one or the other.  Any pixel on this raster have two possible values.

For wave to land, it makes sense to either interpolate, or just create the blank raster, generate an array of points from the pixels, and do the distance calculation on all the pixels themselves giving a complete set.  Same could be said for L2G.


Former-commit-id: 3c383645fd83e68e85241000fd6d007ae159fd5d