Update issue 317
Wrote operation definition to pass into vectorizeRasters along with needed rasters.  However, there are many errors, the big issue being operations done with numpy arrays.  Need to rethink this algorithm and how operation definition should function in this case as opposed to how it is run in the old model.


Former-commit-id: 9d3850c60dd86162be91a20cfdc5bbad2df39c47