Update issue 317
I had to modify the calculations from the old model to better suit how I was approaching the problem.  They do all of their calculations off of one shapefile that has all the information stored in fields of a point that represents a wave farm location.  They then create a raster following.  I am attempting to create rasters first, which contain all the information I need, then use vectorizeRasters to run my calculations and give back the NPV raster.

Added an if,else statement in loop that runs over the '25' years.  If it is the first 'year' then the annualRevenue is '0' and annualCost is equal to IC (installCost + mooringCost + transCost).  Else, the NPV is replaced by addings itself to annualRev - annualCost times the discount rate.


Former-commit-id: 61f2f0f0e08bdf2fa9491ececb59ed23ea1096f7