Update issue 317
Tinkering with the NPV calculations.  Instead of totaling up the capWE for the years specified in the beginning, I'm going to attempt to let the loop that ranges over the years do it for me.  We shall see if my results get me closer to the expected outputs without the outlying anomolies.


Former-commit-id: c023a265efc542bc658883d952da0a07102cae2a