Update issue 340
I was able to index through the DEM raster file and place the correct depth measurement into the corresponding point shape as a field.

The only reason this issue isn't resolved is because the code needs to be placed in a better location and it needs to be decided whether biophysical should do this part, as a similar approach with calculating the wave power and wave energy could be taken.


Former-commit-id: e88afb61d8883b932825943f6af3080827024760