Update issue 340
It seems to make more sense to do all of our calculations (wave power, wave energy) by using the fields from one shapefile, rather than burning a bunch of rasters with a shape layer and interpolating them, and then doing calculations.

Thus I will be remodeling waveEnergy_core.biophysical and add this logic up there.  Once this is done it will no longer need to be in valuation, since valuation reads in that shapefile which will already have the DEM Depth as a field.


Former-commit-id: cc533b82f1dfa73e2d8e0f72b1efae44bb8e8dc7