Update issue 340
I thought I needed to do a geoTransform coordinate system but it turns out that, since I am doing this conversion in biophysical everything is nicely in degrees and lat/long!!

I have added the code for adding the depth the points, and it is verified as correct!

Next step is to blow away most of old code and now do all the calculations on the fields of shape file.  Then create a new field with the NPV values.  Then get the points/values and interpolate onto raster.


Former-commit-id: 83f20939560f187cceb16f874cad831423bf1a2a