Update issue 340
Created a temporary function wavePowerShape, which is a copy of wavePower, and altered it to compute wave power based on the fields from the points in the shapefile.  Currently, all the code looks correct, but it does not seem to be adding the wave power field to the shapefile or providing any output.

I get no errors and everything seems sound....


Former-commit-id: e5ab619aa6f8de1ddef28ba83b773bae950c8c49