Update issue 340
I was able to get the field wave power to output to the shapefile with the same value as previously devised from rasters at the certain points.

Note:  I really did not change anything to make this happen.  It was same code as before, so I am not sure what triggered to make it work this time??????

Former-commit-id: 47b1e22c530b15d6dd2223cf261dc723abe28eea