Update issue 340
I was able to generate a wave power raster from the shapefile wave power field and it seems to match up nicely with the previous rasters!  It is obviously slightly different but by an extremely low percentage.  The only major difference I see is that near the shore the values can be off by 50% or more, mainly because the readings of the Depth from the DEM seem to vary.  I don't think it is a big concern as of right now.

Next step is to remove all the old not needed code!


Former-commit-id: 1bad79bd7afc3b6257ee1b8cff802f7aac84e0bf