Update issue 317
Beginning to use new method to calculate NPV.  Instead of rasterizing some of the shape fields, we are instead going to set all the variables we need into the point shapefile and do calculations from those.  Once the calculations are done and we have added the NPV as its own field into the points we will interpolate them onto a raster for output.

Right now I have hopefully completed the adding value to shapefile part, and just need to do the raster part.


Former-commit-id: 950a39415f804b72eb2f75505edf537fe03bc78d