Update issue 317
I have computed the NPV from using the fields in the shapefile and have developed an output raster that interpolates those values.  The output raster seems to be very close to the other outputs, however, it seems to be in a different projection.  The one from the old model looks to be in meters or WGS_1984_UTM_Zone_10N.  The one I am generating is in Degrees or GCS_WGS_1984.  I will need to reproject this to match the earlier output.


Former-commit-id: 414d8a446e6b3822143a7e7f827b7227599daed8