Update issue 317
I was able to reproject the NPV raster from lat/long to meters.  The reason I couldn't or I should say didn't just create the raster in meters to begin with was because the pixel sizes I have in creating the raster are in degrees measurement.  Although I suppose I could find an equation to convert.  It really does not matter at this point, besides a few lines of code.

I used gdal.AutoCreateWarpedVRT() to transform the projection, this returns a copy the re-projected dataset but does not save it, so I had to do a gdal.CreateCopy to save it as a file.

The question remains of what the pixel size / resolution should be for this raster.


Former-commit-id: c47b2e65c325cd03542a66301605591a843f9670