Update issue 271
Wrote test for the function getPoints, which is used in getting geometry from points to use in distance calculations for valuation.  I created my own shapefile with points and geometries that i hand set.  I then run the function on the shapefile and check to make sure the geometries I get back match.


Former-commit-id: cf1602c92d5522027799251e8af68c054d55672e