Update issue 271
Added onto the changeProjection test case.  Now check the spatial reference of the returned shape to make sure its units are in meters and projection is correct at 'WGS_1984_UTM_Zone_10N'

I also check to make sure all the features and fields match so that nothing was accidentally modified.

Only thing to otherwise check would be the geometry, to see if it is indeed in meters and if those meter units match the conversion to degrees.


Former-commit-id: 162ce020803de627667c046ca7b032c29f9f8d5f