Update issue 391
I switched out or edited the function get_points_values() to only take a shapefile, and a string.  It now iterates through all the features retreiving its X and Y position from its geometry as well as the field's value, which is specified by the string.  It then still creates two lists, one of the points and one of the value.

However, this spawned an issue.  This function is used to get the points for the projected shapefile ,which is used to help interpolate over a new raster.  The raster is projected in lat/long, the projected shapefile is in meters.  This is an issue.


Former-commit-id: 4ded8c7d816260a4d7d6afed27dc1c5485ae3b28