Update issue 360
Adding the ability to request a value from an embedded JSON definition and from
a superUI.

Because this is a preprocessing function, the current approach loads the
target embeddedUI's json config to disk and then recurses through that
definition looking for the target element.  When it finds the target element,
the value is returned as expected.

When contacting a super-ui (the parent of the current embedded ui), the ui
uses the reference to the root ui that is now stored in the object registrar to
search.

Note that both these approaches are currently limited by the fact that they
only communicate with the level immediately above and below.


Former-commit-id: 0c6f6be89ed59e61b79113d048d43a55576e4803