Update issue 388
I had to change a laundry list of things in valuation().  In order to convert from lat/long to meters for creating the landing and grid point shapefiles I was forced to keep a coordinatetransformation, however in order to get the lat/long spatialreference I had to create one as a wellknownGeogCS("WCS84").  This is obviously restricting, and actually truthfully I don't know if this is a great idea.

I was able to remove a bunch of code concerning changing the projection of the shapefile, since it is already in its proper projection.

Then, when creating the raster, I am creating it from the extents of the projected shapefile, so the raster is already projected in meters, however, this means that when setting it's pixel size, I have to make up a value (do not know how to uniformly convert Decimal Degrees to UTM).

I suppose I could reproject the shapefile, with the projection I created earlier, create the raster from that, then reproject the raster.


Former-commit-id: 8acd14ebefede447a291410155a98bb651ed2351