Update issue 394
Moved validation labor over to a new ValidatorThread class, initiated by the
Validator class.  Unfortunately, the UI is still responsible for keeping tabs on
whether the validator thread has finished processing, but I think this is a
necessary cost of having the validation logic happening outside the main thread
of control.


Former-commit-id: dae613109a2b78c4c317eeb6645f2ff05ef9ffa6