Update issue 388
I added a function in invest_cython_core that will get the size of the pixel from lat/long to meters.
With this I moved where I set up my coordinatetransformation, to the top so I could pass it in to the new function.
There are still a lot of commented out code here incase it needs to be dug out again.


Former-commit-id: c1c7c681a94c55955e09974f96355b260404af45