Update issue 390
w_e_val now reads in the economic table in a way that upper/lowercase does not matter for the fields.
When saving values to dictionary it sets the name as the key and makes it lowercase.  So as long as it is spelled correctly
it should be good!  Also added some logger.debug statements into this part.


Former-commit-id: 013ae4c368f001a9dd9d4f0513442781534db29c