Update issue 390
This update corresponds to the previous one with edits to w_e_val.  Simplified how machine_econ dictionary is constructed so that it is simply {Name:Value}.  Thus the code reflects that when retreiving values.  Also, when checking the type of land grid point, I set the type to lowercase and check it against a lowercase string, to take out the case sensitivity issue.


Former-commit-id: 171309521aa4059649ac43574cc86e7e1f84219e