Update issue 390
Working in w_e_bio I was able to provide more efficient code to reading in Machine Performance table and Machine Parameter table.

For machine performance I now specifically get the first row and save it as the column headers, then get the rest of my rows instead of putting and 'if, else' statement in the for loop.

For machine parameter I set the key to the name and made it lowercase, as well as set that key to direct corresponding value.


Former-commit-id: 746fc35ef06a54469ef401a82802eca4a1bb51cf