Update issue 402
I wrote some rough code and configured biophysical so that it will create rasters as in the projection it should be from the start.  The code works, but is a little cludgy and there has to be a more neat, simpler approach.

I specifically put most of my changes into the 'if aoi' 'else' sections.


Former-commit-id: 8c80cffeebe700ed0954d8b6df70dce36ff88c82