Update issue 369
It has been decided on to change the format of the CSV/Excel file and
I have changed it to include the following column headers:
ID    TYPE    LAT    LONG    LOCATION

To match up the landing point with its corresponding grid point, they will share the same
'ID' integer.

With this format in mind I did a rough hack of changing the construction of a useful
dictionary in w_e_valuation and applied it to w_e_core.valuation.  Through the
first test it seems to work well.


Former-commit-id: e472f5e4d19c6b2c48c8122c8cc204005566609b