Update issue 525
Was able to get array from raster band and turn that into a list which was passable to
getPercentiles.  For wave power, I am receiving the right percentile intervals.

I did this by first reading in the values from the raster and flattening then to a 1D array.
I then created a mask  " np.ma.masked_array(array, mask = array ==0)" for the array so
that all NoData values ( 0 ) would be irrevelant.  I then compressed this back into an
ndarray and passed it onto the function getPercentile()

Next step is generating the string classifications from this knowledge and getting
it into raster form.


Former-commit-id: 013ebb872e759783a0c4bef247ecc50f765f902b