Update issue 525
I have created the table and added the relevant rows/columns.  I also generated the correct string output based on the getPercentile functions return list.  The pixel count is currently not calculated and needs to be, which can be done by reading in the raster as an array, then using numpy to generate masks from the certain ranges.

The table seems to save and set to the raster but does not want to display in arcmap.


Former-commit-id: bd639868ff2d041410546a28ead34bf389936f55