Update issue 525
Besides the attribute table and color table I was able to generate the wave power percentile raster.

I did this in two steps.  I first read the wave power raster in as an arrary, flatten it to a 1D array, create a mask to point out the nodata values, and get the percentiles from that list of valid data.  Then, I copy the wave power raster and call vectorize1ArgOp using a function that groups the bands based on the percentiles list.

Still need to do this for captured wave energy and npv rasters.

May want to create a separate function that handles this nicely


Former-commit-id: 8ec1622f1e0f4c8a4d284e1a30401a30956cf3f2