Fixes issue 528
This issue was fixed by Rich who created an example module/file in natcap-dev/gdal_raster_attribute.py

Basically I was using the wrong method when trying to create an attribute table for my raster.  I was attempting to use Gdal.RasterAttributeTable module, when in fact this is best used for setting statistics.  It turns out that a .vat.dbf file needs to be created with the same name as the raster tif file.  So we created a dbf file will the desired fields and values.

It should be noted that ArcGIS automatically colors the raster based on the unique values provided in the attribute table.  So in the future it may be that we need to generate our own ColorTable.

It should also be said that the attribute table only works for whole integers and that I had a raster of floating points which did not work, so I had to convert my output raster to a gdal.GDT_Int32 type.


Former-commit-id: ed52c7202e1fbeba24adfd4871b0e07123e6fc6a