Update issue 525
HUGE update!  I was able to extract the functionality of generating the wave power percentile raster and attribute table so that now it works for both wave power and captured wave energy!  Only have npv left to go!

I now have the following functions to create the output:
	create_percentile_rasters()
	get_percentiles()
	create_percentile_ranges()
	create_attribute_table()

create_percentile_rasters() calls the functions below directly to create all the outputs needed for the raster.


Former-commit-id: 9523cb1a4b362aad0f1ca1da7c4de86b5cb5dd76