Update issue 525
There was some faulty code in 'create_percentile_raster' function.  The operation I was passing to vectorize1ArgOp to group the pixels was also trying to keep a count of the number of pixels in each group.  This count tracking caused a concurrency, synchronization issue where I was getting two of one pixel cell trying to update at the same time.  We took this code out of that op function and below, after the band was set with its groups, we looped over our identifiers and counted the pixels that way.


Former-commit-id: 1c170cfa531fea61573c421d4dff25d69a36e337