Update issue 525
I enhanced the function create_percentile_ranges so that the ranges are not limited to how many percentiles are wanted.  It will now create and return a list of ranges from any percentile list that is NOT EMPTY.  I added an argument to this function 'start_value' that specifes what the first range should begin from.  I placed this variable in all the calls for this function.


Former-commit-id: d935570cfcfdc1a82ceb7c62736be9e63be0dec0