Fixes issue 548
Changed the nodata value for NPV to -100000.  I kept wave power and wave energy the same at 0 because a value 0 or less does not make sense, so it is a good nodata value.  I also went through and added a nodata argument to a couple functions to handle this new change.


Former-commit-id: cc887f89cf8860717b5e2e1165f6c19bb5793841