Update issue 551
I enhanced the structure for storing and transfering the machine performance information.  Instead of creating a 2D array and storing the periods/heights as random elements nested in that array, I now use a dictionary whose keys are specific to what data they are holding.  I changed this in biophysical and then made necessary changes to the core.


Former-commit-id: 25415c85e255b37c91ff2989068eb56b855f94e3