Update issue 551
Enhanced the way the wave watch 3 data is stored in the dictionary, fixing comment 5.

I now store the matrix representation of the bins in an inner dictionary under the key 'bin_matrix'.  I store the period and height arrays in keys, 'periods', 'heights, respectively.

I added comments, docstring, and code fixes in subsequent files to make this change function properly.


Former-commit-id: 267dd879e9d4e94b65b19058e7c4735d0daa5734