Update issue 537
Dealing with this issue of what is the format of a gdal array?  it's (col, row), but there's transposes all over the place and it's a mess.  In the case of case of ls_factor it matters because the bounding box really is defined on cols and rows.  I'm okay ignoring it most of the time but here we set the nrows and ncols specifically off the bounding box and I wanted it to be right.  So all rows and cols variable names are reversed here.  No functional change.


Former-commit-id: 12cd5c8c860a5d063d54e7de0790afe692004bb4