Update issue 537
Fixing a bug that tried to calculate c/p on the lulc raster, but was using the bounding window of the dem which could be a different size.  Lucky to catch that with the sample data.

During this fix I added some probably unnecessary cache flushing, but it makes me feel better to leave it.


Former-commit-id: 7104f01223c7046d7483be081059606e71897968