Helped optimize clip_shape function in w_e_core.
Instead of looping over every single point for all the polygons, I now make a list of all the polygons and check to see if
each point falls within any of them.  This makes it so we only run through each point once.

There still might be a magic OGR function that does clipping/intersecting better.


Former-commit-id: dfefd88847c894b0ec7b0fc2e68e227a956c592c