Fixes issue 564
I deleted an old test case, that tested the function 'vectorizeRasters', with some wave energy files.  The error was from a file that was removed and that is not applicable anymore.  I see no reason to keep this test case, in which case I would pass in another .tif file, because a duplicate test case is directly above it.


Former-commit-id: aaee6d9e49ad6e30864ef4b6898de61ca68dd5a0