Update issue 602
Adding a function to executor.py to dynamically load a python package based on
the heirarchical package name (such as 'invest_natcap.carbon.carbon_core').
This is addressed recursively using the imp module's find_module() and
load_module functions, as suggested by the imp module documentation for
find_module().

I've tested this on my own by attempting to import warp.warp (see the
water funds json config) instead of using the target script and it appears to
work fine.  At the moment, the dynamic module loading in the Executor class
assumes that the executeable function is still called 'execute' and that it
takes a single python dictionary as its argument.

If we wish to dynamically specify the function to be used, we can follow the
pattern suggested in http://stackoverflow.com/a/3071/299084 by using the
getattr() function, as long as we know which function to call.


Former-commit-id: 7cc8b42ef39298b8c2d189cfcfa006faa7dd67a9