Update issue 587
Adding a vertical stretch to the very bottom of the tab layout.  While the dies
address the immediate usability issue, it doesn't address the fact that the
tab's area should not extend beyoud the lower boundary of the window without
enabling the scrollbar ... as it stands, the scrollbar is still there and it
looks like the tab itself is still taking up all the room it posisbly can.

I'll leave this issue open for now, despite the fact that the immediate
usability concern has been addressed.


Former-commit-id: 350f88d41cd068a0f43519fdcc1c4cfe4095d338