Update issue 573
Added comments and a get_mask function to get the mask sub_shed ids in an array.  Thinking of using this instead of rasterizing layer each time I create new mean/area raster.  Changed the function create_volume_raster to create_area_raster because that is what it is actually doing.


Former-commit-id: b94eb1e31953f87f2167bbb7be2a7484aa5b47f4