Update issue 585
Adding a condition to DynamicPrimitive such that a primitive's output value is
not cast to something if its retrieved value is None.

Adding a condition to the Dropdown class such that if there are no elements
(such as if the DBF or CSV file has no fields or if a table file has not been
selected), None is returned.


Former-commit-id: 558ceabc7267837e40c3e2f3960be4d1ef5d8037