Update issue 658
The previous commit to this issue neglected to check whether the executor thread
had been started, so the executor thread wasn't being started
(controller.is_finished() returns true only when its executor has been started
and it finished).  The new approach checks to see if the opDialog's qtimer
object is active as an alternate check.


Former-commit-id: 1fcbeb8a43629b8a7ddd2a3a501f8743d6a2e52e