Fixing an issue with validation.
Previously, there were certain instances where the UI was experiencing a race
condition.  The UI was triggering an element's validation before the element's
previous validation thread had finished.

This commit adds a check to DynamicPrimitive.validate to ensure that the
validator thread is not alive.  It also adds a check to iui_validator.Validator:
when checking to see if a thread has finished, if the validation thread has not
yet been set, the function in question will return True.


Former-commit-id: 441f13a3276fa85e94a998ce238c5688b408e328