Wave Energy changes to deal AOI args case and casting issue
In biophysical module we need to catch if the args['aoi'] is empty and of length
0.  In w_e_core we need to cast fields for setting shapefile property to int and
double.  I did not have this issue when I was developing on my machine earlier
but this issue did pop up on both James and Rich's machine.  The setField
method needed arguments that were specifically cast to int / double.  They may
have been int32 or something before.  Could have to do with different gdal
versions.


Former-commit-id: dc6117a342e83a582be66d14976e5c47945b5a19