Update issue 641
Making some modifications to TableHandler class:
    * changing entries in self.filetypes to be prepended with '.'
    * Updating the use of the handler classes to reflect the new object
       heirarchy change.

Creating a new AbstractTableHandler class that will act as a sort of abstract
interface class for all classes that handle specific tables.

Creating a new open() function for OGRHandler.

Adjusting a number of parameters to reflect the change in object heirarchy
(namely, that the appropriate handler should be created and used to access
 information in the table).


Former-commit-id: 9a7b524854a8bdf7b303e99b4be13e43ec0216b7