Update issue 641
Creating a general-purpose get_map() function for AbstractTableHandler for use
by all it subclasses.

I realized that since we already have a function to get an entire table into a
list of dictionaries, I realized that a function built into the
AbstractTableHandler class could take this list of dictionaries (which
represents the table itself), run it through a comprehension and return the new
dictionary.

This implementation of get_map() now exists in AbstractTableHandler and could be
subclassed by a given Handler class if necessary.


Former-commit-id: 2f15f26f3161ee00920b62f9a673d6907b09e2a6