Update issue 684
Adding a function map_attribute() to pollination_core and a loop to
pollination_core.biophysical() to iterate over all floral and nesting rasters.

map_attribute takes a value map from the fileio table object passed into the
function and creates a nodata-sensitive dictionary for use in a vectorized
function.  If a pixel value cannot be found in the dictionary (which might
happen if the user forgets an entry in the table or if the table happens to have
duplicate keys in the specified key_field), the output raster's nodata value
will be returned.


Former-commit-id: 239d03ef439e1261fc1d925481cbc03e91bbaff1