Update issue 718
In talking with Rich, we decided that if a particular raster is not in the
documentation, it isn't required to produce that raster as an intermediate.
Taking this into account, I'm currently trying to reimplement the nesting and
floral resource mapping functionality as an all-in-one vectorize step (one step
each for floral and nesting resources).


Former-commit-id: 2f9378385caa6301375e74a27bef0b906e8c27f1