Update issue 641
Adding inline comments for the part of fileio.AbstractTableHandler that creates
the masked fieldnames and original fieldnames.

:::TODO:::
I'll need to use the self.orig_fieldnames dictionary in the construction of the
self.table attribute so that we are accessing the correct information from the
original table object and yet saving them to the correct keys in the row
dictionaries.


Former-commit-id: 9e200cd8462a5436fa8c35bb45427c8a271f3994