Update issue 718
Reworking the map_attribute function to produce a map of the whole resource, not
just a single mapped value as it had previously.

Using the sample dataset, this passes both smoke tests in 10s (total, for the
two smoke tests).  This isn't bad, necessarily, but it would be awesome to make
this faster if possible.


Former-commit-id: aace87ef2cea974628d9526e33635f0e87805c0b