Update issue 684
Adding functionality to create the floral resources raster.  The resulting map
looks pretty close to the floral resources map produced by InVEST 2.2, but it
isn't exactly the same.  To fix this, I'll need to take a closer look at the
older pollination code to see where the difference might be or whether the
difference is simply due to some Arc-related algorithm.


Former-commit-id: e7bdabdcad4ec4d31303e8d7d3a89e0b22093811