Update issue 685
Adding some lines to take the supply matrix and convert it to the foraging
abundance index raster (frm_<beename>).  I elected not to write the non-ag
masked raster to disk because it isn't in the user's guide and can be retrieved
by the user if ag classes are not provided, anyways (since when this is the
case, the model assumes that all LULC classes are ag classes).


Former-commit-id: 43d26b2a20198f6b04819d6cca2977878d50afdf