Update issue 685
Grouping operations together in pollination_core.biophysical().

Also, I replaced the old crop_and_blur function with a new one: clip_and_op,
which takes a python callable object as one of its arguments.  This way, the
same clipping and masking that was previously restricted to only a gaussian blur
can now also be applied to any other functionality as well ... as of this
changeset, it's also being used to do per-element multiplication and per-element
addition on two input matrices.


Former-commit-id: 1f71c3246c7df43d8b5766d43b26ee9ad550a705