Update issue 685
I've refactored a bit of the pollination_core.biophysical() to save the
read the foraging total raster once, ensure that its values are set to 0, and
add each foraging index raster to the sum.  Then, once all species have been
processed individually, the summed matrix is then divided by the number of
species to get the mean foraging index across all species.

The frm_avg (average foraging index) matrix is recorded in the documentation
while the frm_tot matrix is not ... because of this, I've updated the API a
little to refer to this matrix as 'foraging_average' instead of
'foraging_total', and this dataset now saves to output/frm_avg.tif


Former-commit-id: 0ed83aa8f520eca2ddba33257f5333a12be1ebde